Update BoundingBox with named ctor idiom
Previous constructor utilized min / max x y values to
construct the bounding box, but led to issues when
shapes more complicated than rectangle could not
directly construct their extent with an initializer list.

To overcome this, named constructors for the bounding
box will be used; these can be called prior to the body
of the individual shape constructors.